Create the Option component
This component takes an option object as a prop.

Assuming the option object has the properties [word:Text, correct:Booleen, ], it implements a DOM element.

The DOM element has a click event listener. If the option is correct, a 'success' event is dispatched, otherwise, the element will turn red for a few seconds.