Create the Options component.
This component takes the current slide and lesson id as properties.

It then uses those props to query the db for the options of the current slide.

it renders those options using the `Option` component  created in
 (c49c095a66ca590d31546fed3078993580d42ee8)

And lastly, if a 'success' event is dispatched from one of the Option components, this component forwards it.