Create the success slide component.
This component takes the current slide object and lesson id as props.

It then queries the db for any sentence examples the slide may have with.

Using an animation timeline, the component slides through two sections. One to celebrate the user's correct answer. And another section to show one of the sentence examples we queried from the db earlier.

Once the animation timeline is over, the component will send the user to the next slide. And if this was the last slide, the component will despatch a 'lessonEnd' event.